Added mutation (Course) Added input type (Course) Fixed input forms and added an exception for non-existent records in the database

+++ cdod/Schema/Mutations/MutationSchool.cs
@@ -1,6 +1,6 @@
 ﻿using cdod.Schema.InputTypes;
 using cdod.Services;
-using cdodDTOs.DTOs;
+using cdods.s;
 
 namespace cdod.Schema.Mutations
 {
@@ -8,12 +8,12 @@ namespace cdod.Schema.Mutations
     public class MutationSchool
     {
         [UseDbContext(typeof(CdodDbContext))]
-        public async Task<SchoolDTO> CreateSchool(SchoolInput schoolForm, [ScopedService] CdodDbContext dbContext)
+        public async Task<School> CreateSchool(SchoolInput schoolForm, [ScopedService] CdodDbContext dbContext)
         {
-            SchoolDTO school = new SchoolDTO()
+            School school = new School()
             {
                 Name = schoolForm.Name,
-                District = schoolForm.District,
+                District = (District)schoolForm.District,
             };
             dbContext.Schools.Add(school);
             await dbContext.SaveChangesAsync();
@@ -21,14 +21,14 @@ namespace cdod.Schema.Mutations
         }
 
         [UseDbContext(typeof(CdodDbContext))]
-        public async Task<SchoolDTO> UpdateSchool(int id, SchoolInput schoolForm, [ScopedService] CdodDbContext dbContext)
+        public async Task<School> UpdateSchool(int id, SchoolInput schoolForm, [ScopedService] CdodDbContext dbContext)
         {
-            SchoolDTO school = new SchoolDTO()
-            {
-                Id = id,
-                Name = schoolForm.Name,
-                District = schoolForm.District,
-            };
+            School? school = dbContext.Schools.FirstOrDefault(s => s.Id == id);
+            if (school == null) throw new Exception("Уазанной школы не существует");
+
+            school.Name = schoolForm.Name ?? school.Name;
+            school.District = schoolForm.District ?? school.District;
+
             dbContext.Schools.Update(school);
             await dbContext.SaveChangesAsync();
             return school;
@@ -37,7 +37,9 @@ namespace cdod.Schema.Mutations
         [UseDbContext(typeof(CdodDbContext))]
         public async Task<bool> DeleteSchool(int schoolId, [ScopedService] CdodDbContext dbContext)
         {
-            SchoolDTO school = new SchoolDTO() { Id = schoolId };
+            School? school = dbContext.Schools.FirstOrDefault(s => s.Id == schoolId);
+            if (school == null) throw new Exception("Уазанной школы не существует");
+
             dbContext.Schools.Remove(school);
             return await dbContext.SaveChangesAsync() > 0;
         }
