+++ cdod/Schema/Mutations/MutationStudent.cs
@@ -1,6 +1,6 @@
 ﻿using cdod.Schema.InputTypes;
 using cdod.Services;
-using cdodDTOs.DTOs;
+using cdods.s;
 
 namespace cdod.Schema.Mutations
 {
@@ -8,17 +8,17 @@ namespace cdod.Schema.Mutations
     public class MutationStudent
     {
         [UseDbContext(typeof(CdodDbContext))]
-        public async Task<StudentDTO> CreateStudent(StudentInput studentForm, [ScopedService] CdodDbContext dbContext)
+        public async Task<Student> CreateStudent(StudentInput studentForm, [ScopedService] CdodDbContext dbContext)
         {
-            StudentDTO student = new StudentDTO()
+            Student student = new Student()
             {
                 FirstName = studentForm.FirstName,
                 LastName = studentForm.LastName,
                 Patronymic = studentForm.Patronymic,
-                BirthDate = studentForm.BirthDate,
+                BirthDate = (DateOnly)studentForm.BirthDate,
                 Descriotion = studentForm.Descriotion,
-                SchoolId = studentForm.SchoolId,
-                ParentId = studentForm.ParentId
+                SchoolId = (int)studentForm.SchoolId,
+                ParentId = (int)studentForm.ParentId
             };
             dbContext.Students.Add(student);
             await dbContext.SaveChangesAsync();
@@ -26,19 +26,17 @@ namespace cdod.Schema.Mutations
         }
 
         [UseDbContext(typeof(CdodDbContext))]
-        public async Task<StudentDTO> UpdateStudent(int studentId, StudentInput studentForm, [ScopedService] CdodDbContext dbContext)
+        public async Task<Student> UpdateStudent(int studentId, StudentInput studentForm, [ScopedService] CdodDbContext dbContext)
         {
-            StudentDTO student = new StudentDTO()
-            {
-                Id = studentId,
-                FirstName = studentForm.FirstName,
-                LastName = studentForm.LastName,
-                Patronymic = studentForm.Patronymic,
-                BirthDate = studentForm.BirthDate,
-                Descriotion = studentForm.Descriotion,
-                SchoolId = studentForm.SchoolId,
-                ParentId = studentForm.ParentId
-            };
+            Student? student = dbContext.Students.FirstOrDefault(s => s.Id == studentId);
+            if (student == null) throw new Exception("Указанного студентва не существует");
+            student.FirstName = studentForm.FirstName ?? student.FirstName;
+            student.LastName = studentForm.LastName ?? student.LastName;
+            student.Patronymic = studentForm.Patronymic ?? student.Patronymic;
+            student.BirthDate = studentForm.BirthDate ?? student.BirthDate;
+            student.Descriotion = studentForm.Descriotion ?? student.Descriotion;
+            student.SchoolId = studentForm.SchoolId ?? student.SchoolId;
+
             dbContext.Students.Update(student);
             await dbContext.SaveChangesAsync();
             return student;
@@ -47,10 +45,8 @@ namespace cdod.Schema.Mutations
         [UseDbContext(typeof(CdodDbContext))]
         public async Task<bool> DeleteStudent(int studentId, [ScopedService] CdodDbContext dbContext)
         {
-            StudentDTO student = new StudentDTO()
-            {
-                Id = studentId
-            };
+            Student? student = dbContext.Students.FirstOrDefault(s => s.Id == studentId);
+            if (student == null) throw new Exception("Указанного студентва не существует");
             dbContext.Students.Remove(student);   
             return await dbContext.SaveChangesAsync() > 0;
         }
@@ -58,9 +54,12 @@ namespace cdod.Schema.Mutations
         [UseDbContext(typeof(CdodDbContext))]
         public async Task<bool> DeleteManyStudents(List<int> studentIds, [ScopedService] CdodDbContext dbContext)
         {
-            List<StudentDTO> students=new List<StudentDTO>();
-            foreach (int studentId in studentIds) students.Add(new StudentDTO() { Id = studentId });
-            //dbContext.Students.RemoveRange(studentIds.Select(s => new StudentDTO() { Id = s }));
+            dbContext.Students.RemoveRange(studentIds.Select(s =>
+            {
+                Student? bufStudent = dbContext.Students.FirstOrDefault(sid => sid.Id == s);
+                if (bufStudent == null) throw new Exception($"Вы указали для удаления несущесствующего ученика ID:{s}");
+                return bufStudent;
+            }));
             return await dbContext.SaveChangesAsync() > 0;
         }
     }
